MDAH - Bug 8352 - Add automatic printing of 'hold to pull' notices
Adds the ability to have a notice printed automatically
after a hold has been placed.

The system has the ability to define a separate printer
for each branchcode, so each library can receive a printed
notice for any holds placed, assuming they have set up the
ability to print over the LAN, WAN, or Internet. Notices are
printed to the pickup library.

Signed-off-by: Alex Arnaud <alex.arnaud@biblibre.com>

Conflicts:
	installer/data/mysql/kohastructure.sql
	installer/data/mysql/updatedatabase.pl