MASSCAT - Bug 10300 - Revert Bug 8781 - Access to Transfer page while "independent branches" is on
This bug needs reverted, this explanation was given to me by a
librarian:

For our group it is useful in order to allow the library to "fix" items
that they receive from other libraries for holds requests that have
either not been transferred properly by the lending library, so the item
was never processed through check in to be placed "in transit," -- there
are frequent enough occasions when a library is filling hold requests
from their queue and the transfer is not completely placed into effect
-- the status will be changed to "in transit" but the "Current Location"
is not updated to the borrowing library.  In these instances, the item
arrives at the borrowing location, they scan it at check in and get the
message "Item must be checked in at home library."

In both of these cases, without the "Transfer" ability, I have to make
the transfer for them.  This is not a workable situation.  It is far too
manual a system if I have to personally transfer items to libraries
myself to correct both of these situations.

Test Plan:
1) As a non-superlibrarian, navigate to circulation-home.pl
2) Note the transfer option is non-existant
3) Apply this patch
4) Refresh circulation-home.pl
5) Note the transfer option is displayed

This reverts commit c101f779e587217ae5ac46ee18aa0292c30aa808.