BWS-PKG - Removing koha-upgrade-schema from post-install tasks
Combining upgrade-schema as part of the apt-get install process causes
problems if, for whatever reason, there is a problem with updating a
particular database. Interrupting the upgrade-schema leaves the package
in a broken state, and you have to start over with apt-get install to
fix it.

Removing this line means that upgrade-schema will have to be run manually
after the code is updated.

modified:   debian/koha-common.postinst