BWS-PKG - Bug 19273 - Confirm hold does not work for in-transit holds
If a hold was transferred to another library and that library
uses the "Confirm" button instead of the "Print and confirm" button,
the hold will remain "in transit" and will not be marked as waiting.

The patron will also not receive notification of the hold since it
never gets marked as waiting.

Test Plan:
1) Place a hold at Library A for pickup at Library B
2) Check the item in at Library A, confirm the transfer to Library B
3) Change your logged in branch to Library B
4) Check the item in at Library B
5) Confirm the hold using the "Confirm" button
6) Note the hold is still marked as in transit
7) Apply this patch
8) Repeat steps 1-5
9) Note the hold is marked correctly as waiting when the "Confirm"
   button is used