ARLINGTON - Bug 24857: Update Database
We have a partner that needs Volume Level Holds.
This is a feature that some ILSs have that allows the creation of volumes at the bibliographic level,
with items being optionally assigned to a given volume. This facilitates the ability to place holds
at a volume level, rather than on a specific item or any available item on a record.

Test Plan:
1) Apply the patches for this bug
2) Run updatedatabase.pl
3) Restart all the things!
4) Enable the new syspref EnableVolumes
5) Browse to detail.pl for a record with items
6) Note the new volumes tab
7) Test creating, editing and deleting volumes
8) On the Holdings tab, select one or more items using the checkboxes
8) Note new selection options for setting a volume for items,
   and for clearing a volume for items.
9) Test adding and clearing the volume set for items

Signed-off-by: Andrew Fuerste-Henry <andrew@bywatersolutions.com>

Signed-off-by: Rebecca  Coert <rcoert@arlingtonva.us>
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>